Se cambio el slideshow para que reciba directamente widgets y los muestre independiente del tipo de imagen o si es un widget como un container
