+++ lib/src/labs/slideshow_page.dart
@@ -1,0 +1,143 @@
+import 'package:designs/src/models/slider_model.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:provider/provider.dart';
+
+class SlideShowPage extends StatelessWidget {
+  const SlideShowPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ChangeNotifierProvider(
+      create: (_) => SliderModel(),
+      child: Scaffold(
+        body: Center(
+          child: Column(
+            children: [
+              Expanded(
+                child: _Slides(),
+              ),
+              _Dots(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _Dots extends StatelessWidget {
+  _Dots();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      height: 70,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          _Dot(index: 0),
+          _Dot(
+            index: 1,
+          ),
+          _Dot(index: 2),
+        ],
+      ),
+    );
+  }
+}
+
+class _Dot extends StatelessWidget {
+  final int? index;
+
+  _Dot({
+    Key? key,
+    this.index,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final pageViewIndex = Provider.of<SliderModel>(context).currentPage;
+    return AnimatedContainer(
+      duration: Duration(milliseconds: 200),
+      width: 12,
+      height: 12,
+      margin: EdgeInsets.symmetric(
+        horizontal: 4,
+      ),
+      decoration: BoxDecoration(
+        color: (pageViewIndex >= index! - 0.5 && pageViewIndex < index! + 0.5)
+            ? Colors.blue
+            : Colors.grey,
+        shape: BoxShape.circle,
+      ),
+    );
+  }
+}
+
+class _Slides extends StatefulWidget {
+  _Slides({Key? key}) : super(key: key);
+
+  @override
+  State<_Slides> createState() => _SlidesState();
+}
+
+class _SlidesState extends State<_Slides> {
+  final pageViewController = PageController();
+
+  @override
+  void initState() {
+    // TODO: implement initState
+    pageViewController.addListener(() {
+      print('Pagina actual: ${pageViewController.page}');
+      //Actualizar el provider, sliderModel:
+      Provider.of<SliderModel>(context, listen: false).currentPage =
+          pageViewController.page!;
+    });
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    pageViewController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: PageView(
+        controller: pageViewController,
+        children: [
+          _Slide(
+            svgPath: 'assets/svgs/slide-1.svg',
+          ),
+          _Slide(
+            svgPath: 'assets/svgs/slide-2.svg',
+          ),
+          _Slide(
+            svgPath: 'assets/svgs/slide-3.svg',
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class _Slide extends StatelessWidget {
+  final String svgPath;
+
+  _Slide({
+    required this.svgPath,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        height: double.infinity,
+        width: double.infinity,
+        padding: EdgeInsets.all(30),
+        child: SvgPicture.asset(svgPath));
+  }
+}
